+++ ouds_core/lib/l10n/gen/ouds_localizations_en.dart
@@ -99,6 +99,19 @@ class OudsLocalizationsEn extends OudsLocalizations {
   String get core_textInput_trait_a11y => 'TextField';
 
   @override
+  String get core_phoneNumberInput_a11y => 'Phone Number';
+
+  @override
+  String get core_phoneNumberInput_countrySelector_a11y => 'Country Selector';
+
+  @override
+  String get core_phoneNumberInput_countrySelector_hint_a11y =>
+      'Double tap to select country';
+
+  @override
+  String get core_phoneNumberInput_hint_a11y => 'Double tap to edit';
+
+  @override
   String get core_password_input_hidden_a11y => 'Password hidden';
 
   @override